Move location Search handler into folder
