+++ src/SFA.DAS.Location.Application/Location/Queries/SearchLocations/GetLocationsQuery.cs
@@ -1,6 +1,6 @@
 using MediatR;
 
-namespace SFA.DAS.Location.Application.Location.Queries
+namespace SFA.DAS.Location.Application.Location.Queries.SearchLocations
 {
     public class GetLocationsQuery : IRequest<GetLocationsQueryResult>
     {
